requirements.txt
pandas
numpy
tqdm
matplotlib
seaborn
scipy
scikit-learn
umap-learn
hdbscan
optuna
sentence-transformers
spacy
nltk
gensim
bertopic
colorcet
